Move chart.js functions to new file
